# docs/api/swagger.yaml
openapi: 3.0.0
info:
    title: Contao Manager API
    version: "1.3.0"
paths:
...
    /api/server/config:
        get:
            summary: Gets server configuration
            tags: [Server Configuration]
            responses:
                200:
                    description: The configured server configuration if available, otherwise the detected values.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Server/Config'
        put:
            summary: Sets server configuration
            tags: [Server Configuration]
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                php_cli:
                                    type: string
                                    description: Path to the PHP command line.
                                cloud:
                                    type: boolean
                                    description: Enables or disables use of the Composer Resolver Cloud.

            responses:
                200:
                    description: If the config was successfully saved.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Server/Config'
                400:
                    description: Error message if the PHP CLI is invalid
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ApiProblem'
...
    /api/server/contao:
        get:
            summary: Gets Contao configuration
            tags: [Server Configuration]
            responses:
                200:
                    description: Gets information about the current Contao installation.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Server/Contao'
                500:
                    description: If any of the commands did not run successfully.
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ApiProblem'
                503:
                    description: If the server config is not ready or the command line binary cannot be found.

        post:
            summary: Create the Contao directory structure.
            description: Creates the necessary directory to install Contao. If the current directory is not empty,
                pass a new directory name to be created. If the current directory is empty but not a web dir, pass
                an empty directory to create the `web` directory only.
            tags: [Server Configuration]
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                directory:
                                    type: string|null
                                    description: The directory name or null if only the `web` directory must be created.
            responses:
                201:
                    description: Operation successful
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Server/Contao'
                403:
                    description: Target directory already exists.
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ApiProblem'
                501:
                    description: Operation is not supported (e.g. not running as Phar).


    #############
    #   Users   #
    #############
...
components:
    schemas:
...
        ApiProblem:
            type: object
            properties:
                title:
                    type: string
                type:
                    type: string
                status:
                    type: integer
                detail:
                    type: string
    securitySchemes:
        CookieAuthentication:
            type: apiKey
            description: Uses an authentication cookie called "contao_manager_auth". The cookie is set by the server as HTTP-only after successful login.
            name: contao_manager_auth
            in: cookie
        TokenAuthentication:
            type: apiKey
            description: 'Uses authentication with a user token. Alternatively, token authentication also works using the
                `Authorization: Bearer …` header. The Contao Manager supports a custom header because bearer
                authentication does not work on all server setups.'
            name: Contao-Manager-Auth
            in: header
security:
    - { CookieAuthentication: [] }
    - { TokenAuthentication: [] }